Add Busqueda por id prices

+++ comercio-electronico/src/main/java/com/comercio/electronico/dto/PriceDTO.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @Data
 public class PriceDTO {
 
-    private Long priceList;
+    private Integer priceList;
     private String name;
     private Date startDate;
     private Date endDate;
